♻️ Refactor: add place modal to show all place data
