+++ src/components/MyRouteContent/AddNearbyPlaceModal.tsx/ModalPlaceList.tsx
@@ -23,7 +23,12 @@ const ModalPlaceList = ({ data, onClick, selectedPlace, className }: ModalPlaceL
             {item.mainImg && (
               <>
                 <div className="w-87 h-87 relative flex-shrink-0 overflow-hidden">
-                  <Image src={item.mainImg} alt={item.name} fill className="rounded-m object-cover object-center" />
+                  <Image
+                    src={item.mainImg.replace(/'/g, '')}
+                    alt={item.name}
+                    fill
+                    className="rounded-m object-cover object-center"
+                  />
                 </div>
                 <div className={`${selectedPlace === item.name && 'text-blue'}`}>
                   <h1 className="font-bold">{item.name}</h1>
