+++ src/components/MyRouteContent/AddNearbyPlaceModal.tsx/index.tsx
@@ -1,4 +1,4 @@
-import ModalPlaceList from '../AddPlaceModal.tsx/ModalPlaceList';
+import ModalPlaceList from './ModalPlaceList';
 import { Map, CustomOverlayMap, MapMarker } from 'react-kakao-maps-sdk';
 import { useEffect, useState } from 'react';
 import instance from '@/src/api/axios';
